requirements.txt
Flask
requests
loguru
python-dotenv
python-dotenv